+++ parsers/discogsartistparser.py
